Added unit tests, protectedRoute for App
