+++ movie_planr/src/components/Planner.jsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import SearchForm from './SearchForm';
 import MoviePlace from './MoviePlace';
-import { zip } from '../utils';
+import { zip } from '../utils/utils';
 
 const Planner = ({ places, legs, addtoPlaces, removePlace, setPlannerRef }) => {
-    
+
     const placesWithLegs = zip(places, [null, ...legs]);
     const moviePlaces = placesWithLegs.map(([place, leg]) => <MoviePlace
         key={place.id}
